refactor: pull modals state into its own action/reducer
