+++ src/client/actions.ts
@@ -7,6 +7,7 @@ import * as organizationJoinActions from "./actions/organizationJoin";
 import * as organizationProjectActions from "./actions/organizationProjects";
 import * as projectDataActions from "./actions/projectData";
 import * as projectOptionsActions from "./actions/projectOptions";
+import * as projectModalsActions from "./actions/projectModals";
 import * as projectsActions from "./actions/projects";
 import * as regionConfigActions from "./actions/regionConfig";
 import * as rootActions from "./actions/root";
@@ -19,6 +20,7 @@ export type OrganizationJoinAction = ActionType<typeof organizationJoinActions>;
 export type OrganizationProjectsAction = ActionType<typeof organizationProjectActions>;
 export type ProjectDataAction = ActionType<typeof projectDataActions>;
 export type ProjectOptionsAction = ActionType<typeof projectOptionsActions>;
+export type ProjectModalsAction = ActionType<typeof projectModalsActions>;
 export type ProjectsAction = ActionType<typeof projectsActions>;
 export type RegionConfigAction = ActionType<typeof regionConfigActions>;
 export type RootAction = ActionType<typeof rootActions>;
@@ -32,6 +34,7 @@ export type Action =
   | OrganizationJoinAction
   | ProjectDataAction
   | ProjectOptionsAction
+  | ProjectModalsAction
   | ProjectsAction
   | RegionConfigAction
   | RootAction
