+++ src/client/reducers.ts
@@ -30,11 +30,16 @@ import projectOptionsReducer, {
   ProjectOptionsState,
   initialProjectOptionsState
 } from "./reducers/projectOptions";
+import projectModalsReducer, {
+  ProjectModalsState,
+  initialProjectModalsState
+} from "./reducers/projectModals";
 
 export interface State {
   readonly auth: AuthState;
   readonly project: ProjectState;
   readonly projectOptions: ProjectOptionsState;
+  readonly projectModals: ProjectModalsState;
   readonly projects: ProjectsState;
   readonly organization: OrganizationState;
   readonly organizationJoin: OrganizationJoinState;
@@ -50,6 +55,7 @@ export const initialState: State = {
   organizationProjects: intialOrganizationProjectsState,
   project: initialProjectState,
   projectOptions: initialProjectOptionsState,
+  projectModals: initialProjectModalsState,
   projects: initialProjectsState,
   regionConfig: initialRegionConfigState,
   user: initialUserState
@@ -62,6 +68,7 @@ const allReducers = combineReducers({
   organizationProjects: organizationProjectsReducer,
   project: projectReducer,
   projectOptions: projectOptionsReducer,
+  projectModals: projectModalsReducer,
   projects: projectsReducer,
   regionConfig: regionConfigReducer,
   user: userReducer
